JWT authetication and auth added

+++ controllers/api/v1/post_api.js
@@ -32,17 +32,28 @@ catch(err){
 
 module.exports.destroy = async function(req,res){
     try{
-    let post = await Post.findById(req.params.id);
-        //.id means converting the id into String
-        // if(post.user==req.user.id){
+        let post = await Post.findById(req.params.id);
+
+        if (post.user == req.user.id){
             post.remove();
-            await Comment.deleteMany({ post : req.params.id});
+
+            await Comment.deleteMany({post: req.params.id});
+
+
+    
             return res.json(200, {
-                message:"Post and related Data Deleted"
-            })
+                message: "Post and associated comments deleted successfully!"
+            });
+        }else{
+            return res.json(401, {
+                message: "You cannot delete this post!"
+            });
+        }
     }
     catch(err){
-        console.log("Error" , err);
-        return;
+        console.log('********', err);
+        return res.json(500, {
+            message: "Internal Server Error"
+        });
         }
     }
